+++ R/idf.R
@@ -488,6 +488,9 @@ Idf <- R6::R6Class(classname = "Idf",
         #' contains valid object IDs in current `Idf` object.
         #'
         #' @param id An integer vector to check.
+        #' @param class A single string indicates the class where the objects to
+        #'        check against. If `NULL`, all classes in current `Idf` are
+        #'        used. Default: `NULL`.
         #'
         #' @return A logical vector with the same length as input integer
         #' vector.
@@ -497,8 +500,8 @@ Idf <- R6::R6Class(classname = "Idf",
         #' idf$is_valid_id(c(51, 1000))
         #' }
         #'
-        is_valid_id = function (id)
-            idf_is_valid_object_id(self, private, id),
+        is_valid_id = function (id, class = NULL)
+            idf_is_valid_object_id(self, private, id, class),
         # }}}
 
         # is_valid_name {{{
@@ -514,6 +517,9 @@ Idf <- R6::R6Class(classname = "Idf",
         #' in all methods that take object name(s) as input.
         #'
         #' @param name A character vector to check.
+        #' @param class A single string indicates the class where the objects to
+        #'        check against. If `NULL`, all classes in current `Idf` are
+        #'        used. Default: `NULL`.
         #'
         #' @return A logical vector with the same length as input character
         #' vector.
@@ -526,8 +532,8 @@ Idf <- R6::R6Class(classname = "Idf",
         #' idf$is_valid_name(c("simple one zone (wireframe dxf)", "zone one", "a"))
         #' }
         #'
-        is_valid_name = function (name)
-            idf_is_valid_object_name(self, private, name),
+        is_valid_name = function (name, class = NULL)
+            idf_is_valid_object_name(self, private, name, class),
         # }}}
 
         # object {{{
@@ -2724,16 +2730,22 @@ idf_is_valid_class_name <- function (self, private, class, all = FALSE) {
 # }}}
 # idf_is_valid_object_id {{{
 #' @importFrom checkmate assert_integerish
-idf_is_valid_object_id <- function (self, private, id) {
+idf_is_valid_object_id <- function (self, private, id, class = NULL) {
     assert_valid_type(id, "Object ID", type = "id")
+    assert_string(class, null.ok = TRUE)
     id <- assert_integerish(id, any.missing = FALSE, coerce = TRUE)
-    id %in% idf_object_id(self, private, NULL, simplify = TRUE)
+    id %in% idf_object_id(self, private, class, simplify = TRUE)
 }
 # }}}
 # idf_is_valid_object_name {{{
-idf_is_valid_object_name <- function (self, private, name) {
+idf_is_valid_object_name <- function (self, private, name, class = NULL) {
     assert_valid_type(name, "Object Name", type = "name")
-    stri_trans_tolower(name) %chin% private$idf_env()$object[!is.na(object_name), object_name_lower]
+    assert_string(class, null.ok = TRUE)
+    if (is.null(class)) {
+        stri_trans_tolower(name) %chin% private$idf_env()$object[!is.na(object_name), object_name_lower]
+    } else {
+        stri_trans_tolower(name) %chin% stri_trans_tolower(get_idf_object_name(private$idd_env(), private$idf_env(), class, simplify = TRUE))
+    }
 }
 # }}}
 # idf_is_unsaved {{{