+++ R/transition.R
@@ -389,7 +389,7 @@ trans_funs$f720t800 <- function (idf) {
                     "Results may be different than previous. ",
                     "See InputOutputReference document for details."
                 ),
-                "warning_trans_720_800"
+                "trans_720_800"
             )
         }
 
@@ -652,7 +652,7 @@ trans_funs$f810t820 <- function (idf) {
                 warn(paste0("Invalid 'Fan Control Type' value for object [ID:", id, "] ",
                     "in class 'ZoneHVAC:UnitHeater' in original v8.1 IDF. ",
                     "Expected 'OnOff' or 'Continuous'. Assuming 'OnOff'..."),
-                    "warning_trans_810_820"
+                    "trans_810_820"
                 )
                 fantype[!fantype %in% c("No", "Yes")] <- "No"
             }
@@ -1402,7 +1402,7 @@ trans_funs$f850t860 <- function (idf) {
                             value[[1L]], "' referenced in class ",
                             "'MaterialProperty:MoisturePenetrationDepth:Settings'."
                         ),
-                        "warning_tran_850_860"
+                        "tran_850_860"
                     )
                     den <- NA_real_
                 }
@@ -1630,7 +1630,7 @@ trans_funs$f870t880 <- function (idf) {
                             "Please check your inputs to make sure this ",
                             "reflects your foundation."
                         ),
-                        "warning_trans_870_880"
+                        "trans_870_880"
                     )
 
                     dt
@@ -1672,7 +1672,7 @@ trans_funs$f870t880 <- function (idf) {
                             "Please check your inputs to make sure this ",
                             "reflects your foundation."
                         ),
-                        "warning_trans_870_880"
+                        "trans_870_880"
                     )
 
                     dt
@@ -1897,7 +1897,7 @@ trans_funs$f890t900 <- function (idf) {
         )
     )
     if (nrow(dt1) && idf$object_num("OutdoorAir:Mixer") > 1L) {
-        warn("Multiple 'OutdoorAir:Mixer' object found.", "warning_trans_890_900")
+        warn("Multiple 'OutdoorAir:Mixer' object found.", "trans_890_900")
     }
     # }}}
     # 2: AirflowNetwork:Distribution:Component:ReliefAirFlow {{{
@@ -1909,7 +1909,7 @@ trans_funs$f890t900 <- function (idf) {
         )
     )
     if (nrow(dt2) && idf$object_num("OutdoorAir:Mixer") > 1L) {
-        warn("Multiple 'OutdoorAir:Mixer' object found.", "warning_trans_890_900")
+        warn("Multiple 'OutdoorAir:Mixer' object found.", "trans_890_900")
     }
     # }}}
     # 3: Boiler:HotWater {{{
@@ -1934,7 +1934,7 @@ trans_funs$f890t900 <- function (idf) {
                         "Option 'UseWeatherFile' for 'Start Day of Week' has been removed, ",
                         "start week day is set by the input start date."
                     ),
-                    "warning_trans_890_900"
+                    "trans_890_900"
                 )
                 value[usewthrfile] <- NA_character_
             }
@@ -1983,7 +1983,7 @@ trans_funs$f890t900 <- function (idf) {
                                 "Option 'UseWeatherFile' for 'Start Day of Week' has been removed, ",
                                 "start week day is set by the input start date."
                             ),
-                            "warning_trans_890_900"
+                            "trans_890_900"
                         )
                         value[8L] <- NA_character_
                     }
@@ -1997,7 +1997,7 @@ trans_funs$f890t900 <- function (idf) {
                                 "Option 'UseWeatherFile' for 'Start Day of Week' has been removed, ",
                                 "start week day is set by the input start date."
                             ),
-                            "warning_trans_890_900"
+                            "trans_890_900"
                         )
                         value[8L] <- NA_character_
                     }
@@ -2012,7 +2012,7 @@ trans_funs$f890t900 <- function (idf) {
                                     "Option 'UseWeatherFile' for 'Start Day of Week' has been removed, ",
                                     "start week day is set by the input start date."
                                 ),
-                                "warning_trans_890_900"
+                                "trans_890_900"
                             )
                             value[8L] <- NA_character_
                         }
@@ -2040,7 +2040,7 @@ trans_funs$f890t900 <- function (idf) {
                                     "Option 'UseWeatherFile' for 'Start Day of Week' has been removed, ",
                                     "start week day is set by the input start date."
                                 ),
-                                "warning_trans_890_900"
+                                "trans_890_900"
                             )
                             value[8L] <- NA_character_
                             # Sunday
@@ -2053,7 +2053,7 @@ trans_funs$f890t900 <- function (idf) {
                                         "Invalid 'Start Day of Week' (", surround(value[8L]), ") found. ",
                                         "Assuming 'Sunday'."
                                     ),
-                                    "warning_trans_890_900"
+                                    "trans_890_900"
                                 )
                             }
                         }
@@ -2095,7 +2095,7 @@ trans_funs$f890t900 <- function (idf) {
                             "With 'Number of Times Runperiod to be Repeated' being ", num_rep,
                             "the end year will be ", end_year, ", which is not a leap year. ",
                             "The end date will be reset to Feb 28th."
-                        ), "error_trans_890_900")
+                        ), "trans_890_900")
                         value[6L] <- "28"
                     }
 
@@ -2250,7 +2250,7 @@ trans_funs$f890t900 <- function (idf) {
                 collapse(unique(empty[, {ifelse(is.na(name), "", (name))}])),
                 " was not used by any surfaces, so it has been deleted.",
                 collpase = "\n"
-            ), "warning_trans_890_900")
+            ), "trans_890_900")
             dt11 <- dt11[!empty, on = c("id", "name")]
             fene_daylight_zone <- fene_daylight_zone[!empty, on = c("id_ctrl" = "id", "name_ctrl" = "name")]
         }
@@ -2488,7 +2488,7 @@ trans_funs$f910t920 <- function (idf) {
     warn_removed_as_comment <- function (idf, class) {
         warn(paste0("Class '", class, "' has been removed in EnergyPlus v9.2. ",
             "Objects in that class will be listed as comments in the new output file."
-        ), "warning_trans_910_920")
+        ), "trans_910_920")
 
         # separate by objects
         cmt <- idf$to_string(class = class, header = FALSE, format = "new_top")
@@ -2517,7 +2517,7 @@ trans_funs$f910t920 <- function (idf) {
             "file. External files must be converted to the new format and saved ",
             "to CSV with a name suffix '-New':\n",
             paste0(obj, ": ", surround(files))
-        ), "warning_trans_910_920")
+        ), "trans_910_920")
 
         # convert
         if (is.null(ascending)) {
@@ -3218,7 +3218,7 @@ trans_funs$f920t930 <- function (idf) {
                         "is not an 'AirflowNetwork:Distribution:Component:Duct' object. ",
                         "No new duct will be added."),
                         value[4L], name[1L], .BY$id
-                    ))
+                    ), "trans_920_930")
                 }
                 val
             }]
@@ -3466,7 +3466,7 @@ trans_funs$f930t940 <- function (idf) {
                 "Field '%s' for object [id:%s] in class 'Output:DebuggingData' ",
                 "is not a number, defaulting to 'No'."),
                 field[1L], .BY$id
-            ))
+            ), "trans_930_940")
             list("No", 0.0)
         }]
 
@@ -3483,7 +3483,7 @@ trans_funs$f930t940 <- function (idf) {
                 "'Output:Diagnostics' has become an unique-object class in EnergyPlus v9.4. ",
                 "All other objects except the first one found will be listed as comments and ",
                 "their keys will be all consolidated into the first one."
-            ))
+            ), "trans_930_940")
             id_cmt <- unique(dt4$id)[-1L]
             id_left <- dt4$id[1L]
             cmt <- idf$to_string(id_cmt, header = FALSE, format = "new_bot")
@@ -3534,11 +3534,330 @@ trans_funs$f930t940 <- function (idf) {
         warn(paste0("Objects in class 'PythonPlugin:Instance' found. ",
             "Note that the API has been changed from v9.3 and v9.4. ",
             "Please check the docs and update with new state argument."
-        ))
+        ), "trans_930_940")
     }
     # }}}
 
     trans_process(new_idf, idf, rbindlist(mget(paste0("dt", 1:9))))
+
+    trans_postprocess(new_idf, idf$version(), new_idf$version())
+}
+# }}}
+# trans_940_950 {{{
+#' @importFrom checkmate assert_true
+trans_funs$f940t950 <- function (idf) {
+    assert_true(idf$version()[, 1:2] == 9.4)
+
+    target_cls <- c(
+        "Construction:AirBoundary",                    # 1
+        "Coil:Cooling:WaterToAirHeatPump:EquationFit", # 2
+        "Coil:Heating:WaterToAirHeatPump:EquationFit", # 3
+        "Construction:InternalSource",                 # 4
+        "HeatPump:WaterToWater:EquationFit:Cooling",   # 5
+        "HeatPump:WaterToWater:EquationFit:Heating",   # 6
+        "ZoneAirMassFlowConservation",                 # 7
+        "ZoneHVAC:LowTemperatureRadiant:VariableFlow", # 8
+        "ZoneHVAC:LowTemperatureRadiant:ConstantFlow", # 9
+        "ZoneHVAC:Baseboard:RadiantConvective:Water",  # 10
+        "ZoneHVAC:Baseboard:RadiantConvective:Steam"   # 11
+    )
+
+    new_idf <- trans_preprocess(idf, 9.5, target_cls)
+    NUM_ADDED <- 0L
+
+    # 1: Construction:AirBoundary {{{
+    dt1 <- trans_action(idf, "Construction:AirBoundary", min_fields = 6)
+    if (nrow(dt1)) {
+        dt1[index == 2L & stri_trans_tolower(value) %chin% "interiorwindow", {
+            warn(sprintf(paste0(
+                "Option 'InteriorWindow' for field 'Solar and Daylighting Method' ",
+                "in class 'Construction:AirBoundary' is no loger valid in Energyplus v9.5. ",
+                "All air boundaries will be modelled using the 'GroupedZones' method. Object(s) below were affected:\n%s"),
+                paste(sprintf(" #%s| Object '%s' [ID: %i]", lpad(seq_along(index), "0"), name, id), collapse = "\n")
+            ), "trans_940_950")
+        }]
+        dt1[index == 3L & stri_trans_tolower(value) %chin% "irtsurface", {
+            warn(sprintf(paste0(
+                "Option 'IRTSurface' for field 'Radiant Exchange Method' ",
+                "in class 'Construction:AirBoundary' is no loger valid in Energyplus v9.5. ",
+                "All air boundaries will be modelled using the 'GroupedZones' method. Object(s) below were affected:\n%s"),
+                paste(sprintf(" #%s| Object '%s' [ID: %i]", lpad(seq_along(index), "0"), name, id), collapse = "\n")
+            ), "trans_940_950")
+        }]
+
+        dt1 <- dt1[!index %in% c(2L, 3L)]
+        dt1[, index := seq_len(.N), by = "id"]
+    }
+    # }}}
+
+    split_curves <- function(dt, pos, name) {
+        if (!nrow(dt)) return(dt)
+
+        num_objs <- length(unique(dt$id))
+        num <- viapply(pos, length)
+        rng <- range(unlist(pos, use.names = FALSE))
+
+        # fields that are retained
+        dt_remain <- dt[index < rng[1L] | index > rng[2L]]
+        # update field index
+        dt_remain[index > rng[2L], index := seq(rng[1L] + length(pos), length.out = .N), by = "id"]
+
+        # curve fields
+        dt_name <- dt[J(viapply(pos, "[[", 1L, use.names = FALSE)), on = "index"]
+        set(dt_name, NULL, "value", paste(rep(name, each = num_objs), rep(seq.int(num_objs), length(pos))))
+        set(dt_name, NULL, "index", rep(seq(min(pos[[1L]]), length.out = length(pos)), each = num_objs))
+
+        # curve objects
+        ind_curve <- unlist(lapply(pos, function(x) c(1L, x)), use.names = FALSE)
+        dt_curve <- dt[J(ind_curve), on = "index"]
+        dt_curve[J(1L), on = "index", value := dt_name$value]
+
+        # update curve object IDs
+        set(dt_curve, NULL, "id", rep(NUM_ADDED - seq.int(num_objs * length(pos)), rep(num + 1L, num_objs)))
+        # update curve object classes
+        set(dt_curve, NULL, "class", rep(rep(names(pos), num + 1L), num_objs))
+        # update curve object indices
+        set(dt_curve, NULL, "index", rep(unlist(lapply(num + 1L, seq.int), use.names = FALSE), num_objs))
+        # create min max table
+        dt_rng <- dt_curve[, by = "id", {
+            class <- class[1L]
+            num <- data.table::fcase(
+                class[1L] == "Curve:QuadLinear", 8L,
+                class[1L] == "Curve:QuintLinear", 10L
+            )
+            index <- seq.int(num)
+            value <- rep(c("-100", "100"), num / 2L)
+            list(name = NA_character_, class, index, field = NA_character_, value)
+        }]
+        dt_curve <- rbindlist(list(dt_curve, dt_rng))
+        dt_curve[, by = "id", index := seq.int(.N)]
+        # update num of new objects added
+        NUM_ADDED <<- min(dt_curve$id)
+
+        rbindlist(list(dt_remain, dt_name, dt_curve))
+    }
+
+    # 2: Coil:Cooling:WaterToAirHeatPump:EquationFit {{{
+    dt2 <- trans_action(idf, "Coil:Cooling:WaterToAirHeatPump:EquationFit", all = TRUE)
+    dt2 <- split_curves(dt2,
+        list("Curve:QuadLinear" = 11:15, "Curve:QuintLinear" = 16:21, "Curve:QuadLinear" = 22:26),
+        c("WAHPCoolCapCurveTot", "WAHPCoolCapCurveSens", "WAHPCoolPowCurve")
+    )
+    # }}}
+    # 3: Coil:Heating:WaterToAirHeatPump:EquationFit {{{
+    dt3 <- trans_action(idf, "Coil:Heating:WaterToAirHeatPump:EquationFit", all = TRUE)
+    dt3 <- split_curves(dt3,
+        list("Curve:QuadLinear" = 10:14, "Curve:QuadLinear" = 15:19),
+        c("WAHPHeatCapCurve", "WAHPHeatPowCurve")
+    )
+    # }}}
+    # 4: Construction:InternalSource {{{
+    dt4 <- trans_action(idf,
+        c("ConstructionProperty:InternalHeatSource" = "Construction:InternalSource"),
+        insert = list(1)
+    )
+    if (nrow(dt4)) {
+        dt4[J(1L), on = "index", value := paste(name, "Heat Source")]
+        # new construction
+        dt41 <- dt4[index == 2L | index >= 8L]
+        set(dt41, NULL, "class", "Construction")
+        dt41[, index := seq.int(.N), by = "id"]
+        set(dt41, NULL, "id", -rleid(dt41$id) + NUM_ADDED)
+        NUM_ADDED <- min(dt41$id)
+
+        dt4 <- rbindlist(list(dt4[index <= 7L], dt41))
+    }
+    # }}}
+    # 5: HeatPump:WaterToWater:EquationFit:Cooling {{{
+    dt5 <- trans_action(idf, "HeatPump:WaterToWater:EquationFit:Cooling", all = TRUE)
+    dt5 <- split_curves(dt5,
+        list("Curve:QuadLinear" = 10:14, "Curve:QuadLinear" = 15:19),
+        c("WWHPCoolCapCurve", "WWHPCoolPowCurve")
+    )
+    # }}}
+    # 6: HeatPump:WaterToWater:EquationFit:Heating {{{
+    dt6 <- trans_action(idf, "HeatPump:WaterToWater:EquationFit:Heating", all = TRUE)
+    dt6 <- split_curves(dt6,
+        list("Curve:QuadLinear" = 10:14, "Curve:QuadLinear" = 15:19),
+        c("WWHPHeatCapCurve", "WWHPHeatPowCurve")
+    )
+    # }}}
+    # 7: ZoneAirMassFlowConservation {{{
+    dt7 <- trans_action(idf, "ZoneAirMassFlowConservation",
+        reset = list(1L, "Yes", "AdjustMixingOnly"),
+        reset = list(1L, "No", "None")
+    )
+    # }}}
+
+    split_radiant <- function(dt, remain, new, remain_since = NULL) {
+        if (!nrow(dt)) return(dt)
+        dt1 <- dt[J(c(remain)), on = "index", nomatch = NULL]
+        if (!is.null(remain_since)) {
+            dt1 <- rbindlist(list(dt1, dt[index >= remain_since]))
+        }
+        dt2 <- dt[J(c(new)), on = "index", nomatch = NULL]
+
+        # construct design name
+        dt1[, index := seq.int(.N), by = "id"]
+        dt1[J(2L), on = "index", value := paste(value, "Design Object")]
+
+        # construct design object
+        set(dt2, NULL, "class", paste0(dt2$class, ":Design"))
+        dt2[, index := seq.int(.N), by = "id"]
+        dt2[J(1L), on = "index", value := paste(value, "Design Object")]
+        dt2[, by = "id", id := NUM_ADDED - .GRP]
+        NUM_ADDED <<- min(dt2$id)
+
+        dt <- rbindlist(list(dt1, dt2))
+    }
+
+    # 8: ZoneHVAC:LowTemperatureRadiant:VariableFlow {{{
+    dt8 <- trans_action(idf, "ZoneHVAC:LowTemperatureRadiant:VariableFlow")
+    dt8 <- split_radiant(dt8,
+        c(1L, 1L, 2:4, 8, 13L, 16:18, 22L, 25:27, 32:33),
+        c(1L, 5:7, 9:12, 14:15, 19:24, 28:31, 34L)
+    )
+    # }}}
+    # 9: ZoneHVAC:LowTemperatureRadiant:ConstantFlow {{{
+    dt9 <- trans_action(idf, "ZoneHVAC:LowTemperatureRadiant:ConstantFlow")
+    dt9 <- split_radiant(dt9,
+        c(1L, 1L, 2:4, 8, 12:15, 18:29, 32:33),
+        c(1L, 5:7, 9:11, 16:17, 30:31)
+    )
+    # }}}
+    # 10: ZoneHVAC:Baseboard:RadiantConvective:Water {{{
+    dt10 <- trans_action(idf, "ZoneHVAC:Baseboard:RadiantConvective:Water")
+    dt10 <- split_radiant(dt10,
+        c(1L, 1L, 2:6, 8, 11),
+        c(1L, 7L, 9:10, 12:14),
+        15L
+    )
+    # }}}
+    # 11: ZoneHVAC:Baseboard:RadiantConvective:Steam {{{
+    dt11 <- trans_action(idf, "ZoneHVAC:Baseboard:RadiantConvective:Steam")
+    dt11 <- split_radiant(dt11,
+        c(1L, 1L, 2:4, 6L, 9:10),
+        c(1L, 5L, 7:8, 11:13),
+        14L
+    )
+    dt11[order(id)]
+    # }}}
+
+    trans_process(new_idf, idf, rbindlist(mget(paste0("dt", 1:11))))
+
+    trans_postprocess(new_idf, idf$version(), new_idf$version())
+}
+# }}}
+# trans_950_960 {{{
+trans_funs$f950t960 <- function (idf) {
+    assert_true(idf$version()[, 1:2] == 9.5)
+
+    target_cls <- c(
+        "AirflowNetwork:MultiZone:ReferenceCrackConditions", # 1
+        "AirLoopHVAC:OutdoorAirSystem",                      # 2
+        "BuildingSurface:Detailed",                          # 3
+        "Ceiling:Adiabatic",                                 # 4
+        "Ceiling:Interzone",                                 # 5
+        "Controller:MechanicalVentilation",                  # 6
+        "Floor:Detailed",                                    # 7
+        "Floor:GroundContact",                               # 8
+        "Floor:Adiabatic",                                   # 9
+        "Floor:Interzone",                                   # 10
+        "GroundHeatExchanger:System",                        # 11
+        "InternalMass",                                      # 12
+        "RoofCeiling:Detailed",                              # 13
+        "Sizing:System",                                     # 14
+        "Roof",                                              # 15
+        "PerformancePrecisionTradeoffs",                     # 16
+        "Wall:Detailed",                                     # 17
+        "Wall:Exterior",                                     # 18
+        "Wall:Adiabatic",                                    # 19
+        "Wall:Underground",                                  # 20
+        "Wall:Interzone"                                     # 21
+    )
+
+    new_idf <- trans_preprocess(idf, 9.6, target_cls)
+
+    # 1: AirflowNetwork:MultiZone:ReferenceCrackConditions {{{
+    dt1 <- trans_action(idf, "AirflowNetwork:MultiZone:ReferenceCrackConditions",
+        reset = list(2, NA_character_, "20.0")
+    )
+    # }}}
+    # 2: AirLoopHVAC:OutdoorAirSystem {{{
+    dt2 <- trans_action(idf, "AirLoopHVAC:OutdoorAirSystem", delete = list(4L))
+    # }}}
+    # 3:BuildingSurface:Detailed {{{
+    dt3 <- trans_action(idf, "BuildingSurface:Detailed", insert = list(5L))
+    # }}}
+    # 4: Construction:InternalSource {{{
+    dt4 <- trans_action(idf, "Ceiling:Adiabatic", insert = list(4L))
+    # }}}
+    # 5: Ceiling:Interzone {{{
+    dt5 <- trans_action(idf, "Ceiling:Interzone", insert = list(4L))
+    # }}}
+    # 6: Controller:MechanicalVentilation {{{
+    dt6 <- trans_action(idf, "Controller:MechanicalVentilation",
+        reset = list(4L, "VentilationRateProcedure", "Standard62.1VentilationRateProcedure")
+    )
+    # }}}
+    # 7: Floor:Detailed {{{
+    dt7 <- trans_action(idf, "Floor:Detailed", insert = list(4L))
+    # }}}
+    # 8: Floor:GroundContact {{{
+    dt8 <- trans_action(idf, "Floor:GroundContact", insert = list(4L))
+    # }}}
+    # 9: Floor:Adiabatic {{{
+    dt9 <- trans_action(idf, "Floor:Adiabatic", insert = list(4L))
+    # }}}
+    # 10: Floor:Interzone {{{
+    dt10 <- trans_action(idf, "Floor:Interzone", insert = list(4L))
+    # }}}
+    # 11: GroundHeatExchanger:System {{{
+    dt11 <- trans_action(idf, "GroundHeatExchanger:System", insert = list(10L))
+    if (nrow(dt11)) {
+        id_target <- dt11[index == 9L & is.na(value), id]
+        if (length(id_target)) {
+            dt11[J(id_target, 10L), on = c("id", "index"), value := "UHFCalc"]
+        }
+    }
+    # }}}
+    # 12: InternalMass {{{
+    dt12 <- trans_action(idf, "InternalMass", insert = list(4L))
+    # }}}
+    # 13: RoofCeiling:Detailed {{{
+    dt13 <- trans_action(idf, "RoofCeiling:Detailed", insert = list(4L))
+    # }}}
+    # 14: Sizing:System {{{
+    dt14 <- trans_action(idf, "Sizing:System",
+        reset = list(27L, "VentilationRateProcedure", "Standard62.1VentilationRateProcedure")
+    )
+    # }}}
+    # 15: Roof {{{
+    dt15 <- trans_action(idf, "Roof", insert = list(4L))
+    # }}}
+    # 16: PerformancePrecisionTradeoffs {{{
+    dt16 <- trans_action(idf, "PerformancePrecisionTradeoffs",
+        reset = list(3L, "Mode07", "Mode08"),
+        reset = list(3L, "Mode06", "Mode07")
+    )
+    # }}}
+    # 17: Wall:Detailed {{{
+    dt17 <- trans_action(idf, "Wall:Detailed", insert = list(4L))
+    # }}}
+    # 18: Wall:Exterior {{{
+    dt18 <- trans_action(idf, "Wall:Exterior", insert = list(4L))
+    # }}}
+    # 19: Wall:Adiabatic {{{
+    dt19 <- trans_action(idf, "Wall:Adiabatic", insert = list(4L))
+    # }}}
+    # 20: Wall:Underground {{{
+    dt20 <- trans_action(idf, "Wall:Underground", insert = list(4L))
+    # }}}
+    # 21: Wall:Interzone {{{
+    dt21 <- trans_action(idf, "Wall:Interzone", insert = list(4L))
+    # }}}
+
+    trans_process(new_idf, idf, rbindlist(mget(paste0("dt", 1:21))))
 
     trans_postprocess(new_idf, idf$version(), new_idf$version())
 }