+++ tests/testthat/test-transition.R
@@ -2074,3 +2074,385 @@ test_that("Transition v9.3 --> v9.4", {
     )
 })
 # }}}
+# v9.4 --> v9.5 {{{
+test_that("Transition v9.4 --> v9.5", {
+    skip_on_cran()
+    from <- 9.4
+    to <- 9.5
+    expect_is(
+        class = "Idf",
+        idfOri <- temp_idf(from,
+            "Construction:AirBoundary" := list(
+                c("CAB1", "CAB2"),
+                c("InteriorWindow", "GroupedZones"),
+                c("GroupedZones", "IRTSurface")
+            ),
+            "Coil:Cooling:WaterToAirHeatPump:EquationFit" = list(
+                "ClgEF1", "WaterInlet", "WaterOutlet", "AirInlet", "AirOutlet",
+                "Autosize", "Autosize", "Autosize", "Autosize", 5.0,
+                1, 2, 3, 4, 5,
+                1, 2, 3, 4, 5, 6,
+                1, 2, 3, 4, 5,
+                0, 0
+            ),
+            "Coil:Heating:WaterToAirHeatPump:EquationFit" := list(
+                "HtgEF1", "WaterInlet", "WaterOutlet", "AirInlet", "AirOutlet",
+                "Autosize", "Autosize", "Autosize", 5.0,
+                1, 2, 3, 4, 5,
+                1, 2, 3, 4, 5
+            ),
+            "Coil:Cooling:WaterToAirHeatPump:EquationFit" = list("ClgEF2"),
+            "Coil:Heating:WaterToAirHeatPump:EquationFit" = list("HtgEF2"),
+            "Construction:InternalSource" = list(
+                "CIS", 2, 2, 1, 0.1, 0, "layer1", "layer2"
+            ),
+            "Construction:InternalSource" = list("CIS2"),
+            "HeatPump:WaterToWater:EquationFit:Cooling" = list(
+                "WWClgEF1", "Inlet1", "Outlet1", "Inlet2", "Outlet2",
+                "Autosize", "Autosize", "Autosize", "Autosize",
+                1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 8.0, 1.0, "Pump"
+            ),
+            "HeatPump:WaterToWater:EquationFit:Heating" = list(
+                "WWHtgEF1", "Inlet1", "Outlet1", "Inlet2", "Outlet2",
+                "Autosize", "Autosize", "Autosize", "Autosize",
+                1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 8.0, 1.0, "Pump"
+            ),
+            "HeatPump:WaterToWater:EquationFit:Cooling" = list("WWClgEF2"),
+            "HeatPump:WaterToWater:EquationFit:Heating" = list("WWHtgEF2"),
+            "ZoneAirMassFlowConservation" = list("Yes"),
+            "ZoneHVAC:LowTemperatureRadiant:VariableFlow" = list(
+                "RadVF1", "Avail", "Zone", "Surface", "ConvectionOnly",
+                0.01, 0.01, "Autosize", 0.35, "MeanAirTemperature", "HalfFlowPower",
+                "HeatingDesignCapacity", "Autosize", NULL, NULL, "Autosize", "Inlet1", "Outlet1", 2.0, "HtgSp",
+                "CoolingDesignCapacity", "Autosize", NULL, NULL, "Autosize", "Inlet2", "Outlet2", 2.0, "ClgSp",
+                "Off", NULL, NULL, NULL
+            ),
+            "ZoneHVAC:LowTemperatureRadiant:VariableFlow" = list("RadVF2"),
+            "ZoneHVAC:LowTemperatureRadiant:ConstantFlow" = list(
+                "RadCF1", "Avail", "Zone", "Surface", "ConvectionOnly",
+                0.01, 0.01, 400, 0.35, "MeanAirTemperature", 0.8, 8e-05, NULL, 8e5, 10, 0.8, 0.1,
+                "Inlet1", "Outlet1", "HighWSch", "LowWSch", "HighCSch", "LowCSch"
+            ),
+            "ZoneHVAC:LowTemperatureRadiant:ConstantFlow" = list("RadCF2"),
+            "ZoneHVAC:Baseboard:RadiantConvective:Water" = list(
+                "BoardC1", "Avial", "Inlet", "Outlet", 80, 0.01, "HeatingDesignCapacity",
+                "Autosize", NULL, NULL, "Autosize", 0.001, 0.3, 0.3,
+                "Surface1", 0.4, "Surface2", 0.2, "Surface3", 0.1
+            ),
+            "ZoneHVAC:Baseboard:RadiantConvective:Water" = list("BoardC2"),
+            "ZoneHVAC:Baseboard:RadiantConvective:Steam" = list(
+                "Steam1", "Avail", "Inlet", "Outlet", "HeatingDesignCapacity",
+                "Autosize", NULL, NULL, 5, "Autosize", 0.001, 0.3, 0.3,
+                "Surface1", 0.4, "Surface2", 0.2, "Surface3", 0.1
+            ),
+            "ZoneHVAC:Baseboard:RadiantConvective:Steam" = list("Steam2"),
+            .all = FALSE
+        )
+    )
+
+    expect_is(idfVU <- version_updater(idfOri, to), "Idf")
+    expect_warning(idfTR <- transition(idfOri, to), "Construction:AirBoundary")
+
+    expect_equal(
+        idfVU$"Construction:AirBoundary"$CAB1$value(),
+        idfTR$"Construction:AirBoundary"$CAB1$value()
+    )
+    expect_equal(
+        idfVU$"Construction:AirBoundary"$CAB2$value(),
+        idfTR$"Construction:AirBoundary"$CAB2$value()
+    )
+
+    expect_equal(
+        idfVU$"Coil:Cooling:WaterToAirHeatPump:EquationFit"$ClgEF1$value(),
+        idfTR$"Coil:Cooling:WaterToAirHeatPump:EquationFit"$ClgEF1$value()
+    )
+    # version updater remove trailing empty fields
+    expect_equal(
+        idfVU$"Coil:Cooling:WaterToAirHeatPump:EquationFit"$ClgEF2$value(),
+        idfTR$"Coil:Cooling:WaterToAirHeatPump:EquationFit"$ClgEF2$value()[1:13]
+    )
+
+    expect_equal(
+        idfVU$"Coil:Heating:WaterToAirHeatPump:EquationFit"$HtgEF1$value(),
+        idfTR$"Coil:Heating:WaterToAirHeatPump:EquationFit"$HtgEF1$value()
+    )
+    # version updater removes trailing empty fields
+    expect_equal(
+        idfVU$"Coil:Heating:WaterToAirHeatPump:EquationFit"$HtgEF2$value(),
+        idfTR$"Coil:Heating:WaterToAirHeatPump:EquationFit"$HtgEF2$value()[1:11]
+    )
+
+    expect_equal(
+        idfVU$"ConstructionProperty:InternalHeatSource"$"CIS Heat Source"$value(),
+        idfTR$"ConstructionProperty:InternalHeatSource"$"CIS Heat Source"$value()
+    )
+    expect_equal(
+        idfVU$"ConstructionProperty:InternalHeatSource"$"CIS2 Heat Source"$value(),
+        idfTR$"ConstructionProperty:InternalHeatSource"$"CIS2 Heat Source"$value()
+    )
+
+    expect_equal(
+        idfVU$"HeatPump:WaterToWater:EquationFit:Cooling"$"WWClgEF1"$value(),
+        idfTR$"HeatPump:WaterToWater:EquationFit:Cooling"$"WWClgEF1"$value()
+    )
+    # version updater removes trailing empty fields
+    expect_equal(
+        idfVU$"HeatPump:WaterToWater:EquationFit:Cooling"$"WWClgEF2"$value(),
+        idfTR$"HeatPump:WaterToWater:EquationFit:Cooling"$"WWClgEF2"$value()[1:11]
+    )
+
+    expect_equal(
+        idfVU$"HeatPump:WaterToWater:EquationFit:Heating"$"WWHtgEF1"$value(),
+        idfTR$"HeatPump:WaterToWater:EquationFit:Heating"$"WWHtgEF1"$value()
+    )
+    # version updater removes trailing empty fields
+    expect_equal(
+        idfVU$"HeatPump:WaterToWater:EquationFit:Heating"$"WWHtgEF2"$value(),
+        idfTR$"HeatPump:WaterToWater:EquationFit:Heating"$"WWHtgEF2"$value()[1:11]
+    )
+
+    expect_equal(
+        idfVU$"ZoneAirMassFlowConservation"$value(),
+        idfTR$"ZoneAirMassFlowConservation"$value()
+    )
+
+    expect_equal(
+        idfVU$"ZoneHVAC:LowTemperatureRadiant:VariableFlow"$RadVF1$value(),
+        idfTR$"ZoneHVAC:LowTemperatureRadiant:VariableFlow"$RadVF1$value()
+    )
+    expect_equal(
+        idfVU$"ZoneHVAC:LowTemperatureRadiant:VariableFlow"$RadVF2$value(),
+        idfTR$"ZoneHVAC:LowTemperatureRadiant:VariableFlow"$RadVF2$value()
+    )
+
+    # version updater fails to keep the "Hydronnic Tubing Length"
+    # see: https://github.com/NREL/EnergyPlus/issues/9171
+    expect_equal(
+        idfVU$"ZoneHVAC:LowTemperatureRadiant:ConstantFlow"$RadCF1$value()[-6],
+        idfTR$"ZoneHVAC:LowTemperatureRadiant:ConstantFlow"$RadCF1$value()[-6]
+    )
+    expect_equal(
+        idfVU$"ZoneHVAC:LowTemperatureRadiant:ConstantFlow"$RadCF2$value()[-6],
+        idfTR$"ZoneHVAC:LowTemperatureRadiant:ConstantFlow"$RadCF2$value()[-6]
+    )
+
+    expect_equal(
+        idfVU$"ZoneHVAC:Baseboard:RadiantConvective:Water"$BoardC1$value(),
+        idfTR$"ZoneHVAC:Baseboard:RadiantConvective:Water"$BoardC1$value()
+    )
+    expect_equal(
+        idfVU$"ZoneHVAC:Baseboard:RadiantConvective:Water"$BoardC2$value(),
+        idfTR$"ZoneHVAC:Baseboard:RadiantConvective:Water"$BoardC2$value()
+    )
+
+    expect_equal(
+        idfVU$"ZoneHVAC:Baseboard:RadiantConvective:Steam"$Steam1$value(),
+        idfTR$"ZoneHVAC:Baseboard:RadiantConvective:Steam"$Steam1$value()
+    )
+    expect_equal(
+        idfVU$"ZoneHVAC:Baseboard:RadiantConvective:Steam"$Steam2$value(),
+        idfTR$"ZoneHVAC:Baseboard:RadiantConvective:Steam"$Steam2$value()
+    )
+})
+# }}}
+# v9.5 --> v9.6 {{{
+test_that("Transition v9.5 --> v9.6", {
+    skip_on_cran()
+
+    # Only install EnergyPlus v9.6 for testing v9.5 to v9.6 transition, because
+    # this version of transition breaks `Wall:Detailed`, `Floor:Detailed`, etc.
+    # See: https://github.com/NREL/EnergyPlus/issues/9172
+    if (!is_avail_eplus(9.6)) install_eplus(9.6, local = TRUE)
+    expect_true(is_avail_eplus(9.6))
+
+    from <- 9.5
+    to <- 9.6
+    expect_is(
+        class = "Idf",
+        idfOri <- temp_idf(from,
+            "AirflowNetwork:MultiZone:ReferenceCrackConditions" = list("AN", NULL),
+            "AirLoopHVAC:OutdoorAirSystem" = list(
+                "OAS", "OAS Ctrl", "OAS Equip", "OAS Avail"
+            ),
+            "BuildingSurface:Detailed" = list(
+                "Surface", "Wall", "Const", "Zone", "Outdoors", NULL,
+                "SunExposed", "WindExposed", 0.5, 3,
+                0, 0, 0, 1, 1, 1, 2, 2, 2
+            ),
+            "Ceiling:Interzone" = list(
+                "CI", "Const", "Zone", "Outdoors", 0, 90, 0, 0, 0, 0.5, 0.5
+            ),
+            "Ceiling:Adiabatic" = list(
+                "CA", "Const", "Zone", 0, 90, 0, 0, 0, 0.5, 0.5
+            ),
+            "Controller:MechanicalVentilation" = list(
+                "CMV", "Avail", "Yes", "VentilationRateProcedure", 1.0, "Zone", "OASpec", "AirDistSpec"
+            ),
+            "Floor:Detailed" = list(
+                "FD", "Const", "Zone", "Outdoors", NULL,
+                "SunExposed", "WindExposed", 0.5, 3,
+                0, 0, 0, 1, 1, 1, 2, 2, 2
+            ),
+            "Floor:GroundContact" = list(
+                "FG", "Const", "Zone", 0, 90, 0, 0, 0, 0.5, 0.5
+            ),
+            "Floor:Adiabatic" = list(
+                "FA", "Const", "Zone", 0, 90, 0, 0, 0, 0.5, 0.5
+            ),
+            "Floor:Interzone" = list(
+                "FI", "Const", "Zone", "Outdoors", 0, 90, 0, 0, 0, 0.5, 0.5
+            ),
+            "GroundHeatExchanger:System" = list(
+                "GHES1", "Inlet", "Outlet", 0.0003,
+                "Site:GroundTemperature:Undisturbed:KusudaAchenbach",
+                "Vertical Ground Heat Exchanger Ground Temps",
+                0.7, 2347000,
+                "Vertical Ground Heat Exchanger g-functions"
+            ),
+            "GroundHeatExchanger:System" = list(
+                "GHES2", "Inlet", "Outlet", 0.0003,
+                "Site:GroundTemperature:Undisturbed:KusudaAchenbach",
+                "Vertical Ground Heat Exchanger Ground Temps",
+                0.7, 2347000, NULL,
+                "Array1", "Single1", "Array2", "Single2"
+            ),
+            "InternalMass" = list("IM", "Const", "Zone", 50),
+            "RoofCeiling:Detailed" = list(
+                "RD", "Const", "Zone", "Outdoors", NULL,
+                "SunExposed", "WindExposed", 0.5, 3,
+                0, 0, 0, 1, 1, 1, 2, 2, 2
+            ),
+            "Sizing:System" = list("SS", ..27 = "VentilationRateProcedure"),
+            "Roof" = list(
+                "CI", "Const", "Zone", 0, 90, 0, 0, 0, 0.5, 0.5
+            ),
+            "PerformancePrecisionTradeoffs" = list("Yes", "ScriptF", "Mode06"),
+            "Wall:Detailed" = list(
+                "WD", "Const", "Zone", "Outdoors", NULL,
+                "SunExposed", "WindExposed", 0.5, 3,
+                0, 0, 0, 1, 1, 1, 2, 2, 2
+            ),
+            "Wall:Exterior" = list(
+                "WE", "Const", "Zone", 0, 90, 0, 0, 0, 0.5, 0.5
+            ),
+            "Wall:Adiabatic" = list(
+                "WA", "Const", "Zone", 0, 90, 0, 0, 0, 0.5, 0.5
+            ),
+            "Wall:Underground" = list(
+                "WU", "Const", "Zone", 0, 90, 0, 0, 0, 0.5, 0.5
+            ),
+            "Wall:Interzone" = list(
+                "WI", "Const", "Zone", "Outdoors", 0, 90, 0, 0, 0, 0.5, 0.5
+            ),
+            .all = FALSE
+        )
+    )
+
+    expect_is(idfVU <- version_updater(idfOri, to), "Idf")
+    expect_is(idfTR <- transition(idfOri, to), "Idf")
+
+    expect_equal(
+        idfVU$"AirflowNetwork:MultiZone:ReferenceCrackConditions"$AN$value(),
+        idfTR$"AirflowNetwork:MultiZone:ReferenceCrackConditions"$AN$value()
+    )
+
+    expect_equal(
+        idfVU$"AirLoopHVAC:OutdoorAirSystem"$OAS$value(),
+        idfTR$"AirLoopHVAC:OutdoorAirSystem"$OAS$value()
+    )
+
+    expect_equal(
+        idfVU$"BuildingSurface:Detailed"$Surface$value(),
+        idfTR$"BuildingSurface:Detailed"$Surface$value()
+    )
+
+    expect_equal(
+        idfVU$"Ceiling:Interzone"$CI$value(),
+        idfTR$"Ceiling:Interzone"$CI$value()
+    )
+
+    expect_equal(
+        idfVU$"Ceiling:Adiabatic"$CA$value(),
+        idfTR$"Ceiling:Adiabatic"$CA$value()
+    )
+
+    expect_equal(
+        idfVU$"Controller:MechanicalVentilation"$CMV$value(),
+        idfTR$"Controller:MechanicalVentilation"$CMV$value()
+    )
+
+    expect_equal(
+        idfVU$"Floor:Detailed"$FD$value(),
+        idfTR$"Floor:Detailed"$FD$value()
+    )
+
+    expect_equal(
+        idfVU$"Floor:GroundContact"$FG$value(),
+        idfTR$"Floor:GroundContact"$FG$value()
+    )
+
+    expect_equal(
+        idfVU$"Floor:Adiabatic"$FA$value(),
+        idfTR$"Floor:Adiabatic"$FA$value()
+    )
+
+    expect_equal(
+        idfVU$"Floor:Interzone"$FI$value(),
+        idfTR$"Floor:Interzone"$FI$value()
+    )
+
+    expect_equal(
+        idfVU$"GroundHeatExchanger:System"$GHES1$value(),
+        idfTR$"GroundHeatExchanger:System"$GHES1$value()
+    )
+    expect_equal(
+        tolower(idfVU$"GroundHeatExchanger:System"$GHES2$value()),
+        tolower(idfTR$"GroundHeatExchanger:System"$GHES2$value())
+    )
+
+    expect_equal(
+        idfVU$"InternalMass"$IM$value(),
+        idfTR$"InternalMass"$IM$value()
+    )
+
+    expect_equal(
+        idfVU$"RoofCeiling:Detailed"$RD$value(),
+        idfTR$"RoofCeiling:Detailed"$RD$value()
+    )
+
+    expect_equal(
+        idfVU$"Sizing:System"[[1]]$value(),
+        idfTR$"Sizing:System"[[1]]$value()
+    )
+
+    expect_equal(
+        idfVU$"PerformancePrecisionTradeoffs"$value(),
+        idfTR$"PerformancePrecisionTradeoffs"$value()
+    )
+
+    expect_equal(
+        idfVU$"Wall:Detailed"$WD$value(),
+        idfTR$"Wall:Detailed"$WD$value()
+    )
+
+    expect_equal(
+        idfVU$"Wall:Exterior"$WE$value(),
+        idfTR$"Wall:Exterior"$WE$value()
+    )
+
+    expect_equal(
+        idfVU$"Wall:Adiabatic"$WA$value(),
+        idfTR$"Wall:Adiabatic"$WA$value()
+    )
+
+    expect_equal(
+        idfVU$"Wall:Underground"$WU$value(),
+        idfTR$"Wall:Underground"$WU$value()
+    )
+
+    expect_equal(
+        idfVU$"Wall:Interzone"$WI$value(),
+        idfTR$"Wall:Interzone"$WI$value()
+    )
+})
+# }}}