+++ tools/data.R
@@ -1,9 +1,9 @@
-for (f in list.files(here::here("tools/R"), "*.R$", full.names = TRUE)) source(f)
+for (f in list.files("tools/R", "*.R$", full.names = TRUE)) source(f)
 
 set_proxy()
 
 eplus_src <- file.path(Sys.getenv("USERPROFILE"), "Dropbox/github_repo/EnergyPlus")
-update_internal_data(eplus_src, force = TRUE)
+update_internal_data(eplus_src)
 
 unset_proxy()